// frontend/src/generated/SearchQuery.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SearchQuery = { root: string, query: string, case_sensitive: boolean, whole_word: boolean, max_results: number, regex: boolean, include_glob: string | null, exclude_glob: string | null, context_lines: number, respect_gitignore: boolean, };