// frontend/src/generated/ReplaceResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ReplaceFileResult } from "./ReplaceFileResult";

export type ReplaceResponse = { operation_id: string, files: Array<ReplaceFileResult>, preview: boolean, cancelled: boolean, };